requirements.txt
google-generativeai
requests
schedule